feat: add /cart POST route with middleware
this middleware validates header and body, the header need to have a
authorization key with a JWT token.
The body needs to have the same keys that /create-new-product needs.
It only validates data in the sense that they are in the same format needed.
On the future validations for token and product will be needed.